+++ solr/core/src/java/org/apache/solr/schema/RandomSortField.java
@@ -133,7 +133,7 @@ public class RandomSortField extends FieldType {
 
             @Override
             public void setTopValue(Integer value) {
-              topVal = value.intValue();
+              topVal = value;
             }
 
             @Override
