+++ solr/solrj/src/java/org/apache/solr/common/util/FastJavaBinDecoder.java
@@ -339,7 +339,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Byte.valueOf((byte) entry.numericVal);
+        return (byte) entry.numericVal;
       }
     },
     _SHORT(SHORT, LOWER_5_BITS, DataEntry.Type.INT) {
@@ -351,7 +351,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Short.valueOf((short) entry.numericVal);
+        return (short) entry.numericVal;
       }
     },
     _DOUBLE(DOUBLE, LOWER_5_BITS, DataEntry.Type.DOUBLE) {
@@ -363,7 +363,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Double.valueOf(entry.doubleVal);
+        return entry.doubleVal;
       }
     },
     _INT(INT, LOWER_5_BITS, DataEntry.Type.INT) {
@@ -374,7 +374,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Integer.valueOf((int) entry.numericVal);
+        return (int) entry.numericVal;
       }
     }, // signed integer
     _LONG(LONG, LOWER_5_BITS, DataEntry.Type.LONG) {
@@ -385,7 +385,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Long.valueOf(entry.numericVal);
+        return entry.numericVal;
       }
     },
     _FLOAT(FLOAT, LOWER_5_BITS, DataEntry.Type.FLOAT) {
@@ -396,7 +396,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Float.valueOf((float) entry.doubleVal);
+        return (float) entry.doubleVal;
       }
     },
     _DATE(DATE, LOWER_5_BITS, DataEntry.Type.DATE) {
